add admin dashboard and upload new pizza

+++ pages/api/products/[id].js
@@ -5,9 +5,11 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
 
   dbConnect();
+  const token = cookies.token;
 
   if (method === 'GET') {
     try {
@@ -18,6 +20,9 @@ export default async function handler(req, res) {
     }
   }
   if (method === 'PUT') {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json('You are not authenticated');
+    }
     try {
       const product = await Product.findByIdAndUpdate(id, req.body, {
         new: true,
@@ -28,6 +33,9 @@ export default async function handler(req, res) {
     }
   }
   if (method === 'DELETE') {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json('You are not authenticated');
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(201).json('The product has been deleted');
